+++ packages/awskit-s3/eio/awskit_s3_eio.ml
@@ -2,6 +2,7 @@ module Runtime = Awskit_eio.Runtime
 module S3 = Awskit_s3.Make (Runtime)
 
 type t = S3.t
+type +'a io = 'a
 
 module File_transfer = Transfer
 module Body_reader = File_transfer.Make_body_reader (Runtime) (S3)